QAmqpPendingFrame: Internal frame type for storage.
Certain AMQP methods are synchronous in nature, thus all other methods
should be queued while the reply from that synchronous operation comes
back.

Since the standard QAmqpFrame class cannot be stored on its own (it is
an abstract class), we create a subclass that grabs the to-be-sent
payload and stashes it away safely and reproduces the original frame on
request.