QAmqpExchange: Spoof the precondition failed on re-declare.
If we re-declare an exchange, spoof the "precondition failed" message
we'd get when things don't match.  This is not allowed in the AMQP
specs, so we know if we were to send it to the server, the server is
going to slam the door on us.

This is incredibly dodgy, but it is how other code will be expecting
this condition.  We do it via a QTimer::oneShot call so that we avoid
racing the QTest framework when testing for this problem.